[ANDROSDK-445] Move D2ErrorCode
