+++ core/src/main/java/org/hisp/dhis/android/core/common/D2CallExecutor.java
@@ -32,6 +32,7 @@ import android.util.Log;
 
 import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
 import org.hisp.dhis.android.core.data.database.Transaction;
+import org.hisp.dhis.android.core.maintenance.D2ErrorCode;
 
 import java.util.ArrayList;
 import java.util.Collection;
