+++ core/src/main/java/org/hisp/dhis/android/core/dataset/DataSetCompleteRegistrationCallFetcher.java
@@ -33,7 +33,7 @@ import android.support.annotation.NonNull;
 import org.hisp.dhis.android.core.calls.fetchers.CallFetcher;
 import org.hisp.dhis.android.core.common.APICallExecutor;
 import org.hisp.dhis.android.core.common.D2CallException;
-import org.hisp.dhis.android.core.common.D2ErrorCode;
+import org.hisp.dhis.android.core.maintenance.D2ErrorCode;
 import org.hisp.dhis.android.core.utils.Utils;
 
 import java.util.ArrayList;
