+++ core/src/main/java/org/hisp/dhis/android/core/maintenance/D2ErrorCode.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.common;
+package org.hisp.dhis.android.core.maintenance;
 
 public enum D2ErrorCode {
     ALREADY_AUTHENTICATED,
