+++ core/src/main/java/org/hisp/dhis/android/core/user/UserAuthenticateCallErrorCatcher.java
@@ -29,7 +29,7 @@
 package org.hisp.dhis.android.core.user;
 
 import org.hisp.dhis.android.core.common.APICallErrorCatcher;
-import org.hisp.dhis.android.core.common.D2ErrorCode;
+import org.hisp.dhis.android.core.maintenance.D2ErrorCode;
 
 import java.io.IOException;
