+++ core/src/sharedTest/java/org/hisp/dhis/android/core/data/maintenance/D2ErrorSamples.java
@@ -28,7 +28,7 @@
 
 package org.hisp.dhis.android.core.data.maintenance;
 
-import org.hisp.dhis.android.core.common.D2ErrorCode;
+import org.hisp.dhis.android.core.maintenance.D2ErrorCode;
 import org.hisp.dhis.android.core.maintenance.D2Error;
 import org.hisp.dhis.android.core.maintenance.D2ErrorComponent;
